+++ packages/integration-sdk-runtime/src/execution/executeIntegration.ts
@@ -6,6 +6,7 @@ import {
   ExecutionHistory,
   IntegrationInstance,
   IntegrationInstanceConfig,
+  IntegrationInstanceExecutionContext,
   IntegrationInvocationConfig,
   IntegrationLogger,
   IntegrationStepResult,
@@ -14,7 +15,6 @@ import {
   Relationship,
   StepExecutionContext,
   StepResultStatus,
-  StepStartStates,
 } from '@jupiterone/integration-sdk-core';
 
 import {
@@ -49,6 +49,7 @@ import { trimStringValues } from './utils/trimStringValues';
 import { validateStepStartStates } from './validation';
 import { processDeclaredTypesDiff } from './utils/processDeclaredTypesDiff';
 import { DuplicateKeyTracker } from './duplicateKeyTracker';
+import { getIngestionSourceStepStartStates } from './utils/getIngestionSourceStepStartStates';
 
 export interface ExecuteIntegrationResult {
   integrationStepResults: IntegrationStepResult[];
@@ -165,7 +166,7 @@ async function tryPublishDiskUsageMetric<
  * using context that was provided.
  */
 export async function executeWithContext<
-  TExecutionContext extends ExecutionContext,
+  TExecutionContext extends IntegrationInstanceExecutionContext,
   TStepExecutionContext extends StepExecutionContext,
 >(
   context: TExecutionContext,
@@ -226,11 +227,19 @@ export async function executeWithContext<
       throw err;
     }
 
-    const stepStartStates: StepStartStates =
+    const configStepStartStates =
       (await config.getStepStartStates?.(context)) ??
       getDefaultStepStartStates(config.integrationSteps);
 
-    validateStepStartStates(config.integrationSteps, stepStartStates);
+    validateStepStartStates(config.integrationSteps, configStepStartStates);
+
+    const stepStartStates = getIngestionSourceStepStartStates({
+      integrationSteps: config.integrationSteps,
+      configStepStartStates,
+      disabledSources: context.instance?.disabledSources?.map(
+        (source) => source.ingestionSourceId,
+      ),
+    });
 
     if (shouldPublishDiskUsageMetric) {
       diskUsagePublishInterval = createDiskUsagePublishInterval();